+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Cancel pending and running child media downloads when removing an offline collection (Offliner)
+
 ## [0.11.17] - 2026-05-04
 
 ### Added